# public/openapi.yaml
openapi: 3.1.0
info:
  title: Mikhnevich Bar API
  version: 1.0.0
  description: >
    REST API винной коллекции. Одна сущность — Wine (вино).
    Тот же бэкенд обслуживает и веб-сайт, и iOS-приложение.

    База данных: PostgreSQL (Neon). Бэкенд: Next.js Route Handlers.

    Все даты приходят в ISO-формате (UTC). `purchaseDate` — только дата (YYYY-MM-DD),
    `drankAt` / `createdAt` / `updatedAt` — дата-время (ISO 8601).

    АВТОРИЗАЦИЯ: все эндпоинты (кроме POST /api/auth/login) требуют входа.
    Каждый пользователь видит и меняет только СВОИ вина; гостевая карта тоже
    приватная. Два способа передать сессию:
      - Веб-сайт: httpOnly-cookie `session` (ставится автоматически после входа).
      - Приложение: заголовок `Authorization: Bearer <token>`, где token получен
        из POST /api/auth/login.
security:
  - bearerAuth: []
  - cookieAuth: []
servers:
  - url: http://localhost:3000
    description: Локальная разработка
  - url: https://ЗАМЕНИТЬ-НА-ПРОДАКШН-ДОМЕН
    description: Продакшн (укажите домен деплоя, например на Vercel)
...
paths:
...
  /api/wines/browse:
    get:
      tags: [Wines]
      summary: Список вин с фильтрами, сортировкой и пагинацией
      description: >
        Основной эндпоинт для UI. По умолчанию отдаёт вина сгруппированными по цвету
        (`sections`). Если передать `flat=1` — отдаёт плоский список с `total/limit/offset`.
      parameters:
        - name: drank
          in: query
          description: "true — раздел «Выпито», иначе активная коллекция"
          schema: { type: boolean, default: false }
        - name: flat
          in: query
          description: "1 — плоский список вместо группировки по цвету"
          schema: { type: string, enum: ["1"] }
        - name: name
          in: query
          description: Поиск по названию или производителю (регистронезависимо)
          schema: { type: string }
        - name: country
          in: query
          description: >
            Фильтр по стране. Можно повторять параметр (?country=France&country=Italy)
            или передать через запятую.
          schema: { type: string }
        - name: region
          in: query
          description: Фильтр по региону (работает, только если выбрана ровно одна страна)
          schema: { type: string }
        - name: priceField
          in: query
          description: По какому ценовому полю фильтровать диапазон цен
          schema:
            type: string
            enum: [purchase, israel, origin, guestBottle, guestGlass]
            default: purchase
        - name: priceMin
          in: query
          schema: { type: number }
        - name: priceMax
          in: query
          schema: { type: number }
        - name: ratingMin
          in: query
          description: >
            Минимальный рейтинг. Для активной коллекции — по vivinoRating,
            для раздела «Выпито» — по drankRating.
          schema: { type: number }
        - name: sortBy
          in: query
          schema:
            type: string
            enum:
              - purchaseDate
              - purchasePrice
              - israelPrice
              - originPrice
              - guestBottlePrice
              - guestGlassPrice
              - vivinoRating
              - collectionValue
              - drankAt
              - drankRating
              - name
              - year
            default: purchasePrice
        - name: sortDir
          in: query
          schema: { type: string, enum: [asc, desc], default: desc }
        - name: limit
          in: query
          description: "Только при flat=1. Размер страницы (1–500, по умолчанию 20)"
          schema: { type: integer, minimum: 1, maximum: 500, default: 20 }
        - name: offset
          in: query
          description: "Только при flat=1. Смещение"
          schema: { type: integer, minimum: 0, default: 0 }
        - name: redLimit
          in: query
          description: "Кол-во красных вин в секции (режим группировки). Аналогично whiteLimit, roseLimit, sparklingLimit"
          schema: { type: integer, minimum: 1, maximum: 500 }
        - name: whiteLimit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500 }
        - name: roseLimit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500 }
        - name: sparklingLimit
          in: query
          schema: { type: integer, minimum: 1, maximum: 500 }
      responses:
        "200":
          description: Результат в режиме секций или плоском режиме (flat=1)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/BrowseSectionsResponse"
                  - $ref: "#/components/schemas/BrowseFlatResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Токен из POST /api/auth/login. Заголовок: Authorization: Bearer <token>"
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: httpOnly-cookie сессии (используется веб-сайтом автоматически)
...
  schemas:
    WineColor:
      type: string
      enum: [red, white, rose, sparkling]
      description: "Цвет вина: red — красное, white — белое, rose — розовое, sparkling — игристое"
    Wine:
      type: object
      description: Вино — так, как оно возвращается из API.
      properties:
        id: { type: string }
        name: { type: string }
        producer: { type: string }
        year:
          type: [string, "null"]
          description: Винтаж (год) строкой
        country: { type: [string, "null"] }
        countryCode:
          type: [string, "null"]
          description: ISO-код страны
        region: { type: [string, "null"] }
        subregion: { type: [string, "null"] }
        grape:
          type: [string, "null"]
          description: Сорт винограда
        ratings:
          type: [string, "null"]
          description: Сырая строка рейтингов (например, из Vivino)
        purchasePrice: { type: [integer, "null"] }
        purchaseCurrency: { type: [string, "null"] }
        originPrice:
          type: [integer, "null"]
          description: Цена на родине
        originCurrency: { type: [string, "null"] }
        israelPrice:
          type: [integer, "null"]
          description: Цена в Израиле
        israelCurrency: { type: [string, "null"] }
        isGuestVisible:
          type: boolean
          description: Показывать ли вино в гостевом меню
        guestBottlePrice:
          type: [integer, "null"]
          description: Цена за бутылку для гостей
        guestGlassPrice:
          type: [integer, "null"]
          description: Цена за бокал для гостей
        purchaseDate:
          type: [string, "null"]
          format: date
          description: Дата покупки (YYYY-MM-DD)
        vivinoRating:
          type: [number, "null"]
          description: Рейтинг Vivino
        quantity:
          type: integer
          description: Количество бутылок в коллекции
        color:
          $ref: "#/components/schemas/WineColor"
        drank:
          type: boolean
          description: Находится ли в разделе «Выпито»
        drankAt:
          type: [string, "null"]
          format: date-time
          description: Когда выпито
        drankRating:
          type: [number, "null"]
          description: Оценка после дегустации (0–10)
        drankNotes: { type: [string, "null"] }
        notes: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        - id
        - name
        - producer
        - isGuestVisible
        - quantity
        - color
        - drank
        - createdAt
        - updatedAt
...
    WineTotals:
      type: object
      properties:
        collection:
          type: integer
          description: Кол-во записей в активной коллекции
        drank:
          type: integer
          description: Кол-во записей в разделе «Выпито»
        bottles:
          type: integer
          description: Суммарное кол-во бутылок в коллекции
        value:
          type: integer
          description: Суммарная стоимость коллекции (по purchasePrice)
      required: [collection, drank, bottles, value]
    BrowseFacets:
      type: object
      properties:
        countries:
          type: array
          items: { type: string }
        regions:
          type: array
          items: { type: string }
    BrowseAppliedFilters:
      type: object
      properties:
        countryKeys:
          type: array
          items: { type: string }
        regionKey: { type: string }
    WineSection:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Wine"
        total:
          type: integer
          description: Всего вин этого цвета под текущими фильтрами
        limit:
          type: integer
      required: [items, total, limit]
    BrowseSectionsResponse:
      type: object
      description: Ответ /api/wines/browse в режиме группировки по цвету.
      properties:
        totals:
          $ref: "#/components/schemas/WineTotals"
        facets:
          $ref: "#/components/schemas/BrowseFacets"
        filters:
          $ref: "#/components/schemas/BrowseAppliedFilters"
        sections:
          type: object
          properties:
            red: { $ref: "#/components/schemas/WineSection" }
            white: { $ref: "#/components/schemas/WineSection" }
            rose: { $ref: "#/components/schemas/WineSection" }
            sparkling: { $ref: "#/components/schemas/WineSection" }
      required: [totals, facets, filters, sections]
    BrowseFlatResponse:
      type: object
      description: Ответ /api/wines/browse при flat=1.
      properties:
        totals:
          $ref: "#/components/schemas/WineTotals"
        facets:
          $ref: "#/components/schemas/BrowseFacets"
        filters:
          $ref: "#/components/schemas/BrowseAppliedFilters"
        items:
          type: array
          items:
            $ref: "#/components/schemas/Wine"
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
      required: [totals, facets, filters, items, total, limit, offset]
...